Moved some classes to new files to fix circular import issues
