+++ fast64_internal/oot/oot_skeleton.py
@@ -1,13 +1,14 @@
 import shutil, copy, mathutils, bpy, math, os, re
+from bpy.utils import register_class, unregister_class
+
+from ..utility import *
+from .oot_utility import *
+from .oot_constants import *
+from .oot_model_classes import *
 
 from ..f3d.f3d_parser import *
 from ..f3d.f3d_writer import *
 from ..f3d.f3d_material import TextureProperty, tmemUsageUI
-from bpy.utils import register_class, unregister_class
-from .oot_constants import *
-from .oot_utility import *
-from .oot_f3d_writer import *
-from ..utility import *
 
 ootEnumBoneType = [
 	("Default", "Default", "Default"),
